Linux Audit Rule Configuration Script
This script will configure your system to send logs to a
remote/centralized syslog server or SIEM.  It will also configure audit
rules that are DoD compliant for RHEL systems (unconfirmed) as well as
configure additional monitoring rules you may want to have in place.
Suggestions on improvements are welcome.